@@ -18,7 +18,7 @@
 
 	
 	<properties>
-		<java.version>19</java.version> 
+		<java.version>21</java.version> 
 	</properties>
 	<dependencies>
 
@@ -46,7 +46,8 @@
 		<dependency>
 			<groupId>org.projectlombok</groupId>
 			<artifactId>lombok</artifactId>
-			<optional>true</optional>
+			<version>1.18.30</version>
+    		<scope>provided</scope>
 		</dependency>
 		<dependency>
 			<groupId>org.springframework.boot</groupId>
