@@ -1,11 +1,11 @@
 package br.com.clsystem.CLSystem.controller;
 
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.authentication.AuthenticationManager;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.AuthenticationException;
 import org.springframework.web.bind.annotation.CrossOrigin;
+import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -15,7 +15,7 @@ import org.springframework.web.bind.annotation.RestController;
 @CrossOrigin(origins="*", maxAge = 3600) //Permitir ser acessado de Qualquer fonte
 @RequestMapping("/api")
 public class Login {
-    ]
+    
     private AuthenticationManager authenticationManager;
 
     public Login(AuthenticationManager authenticationManager){
@@ -23,11 +23,11 @@ public class Login {
     }
 
     @PostMapping("/public/employee/login")
-    public String login(@RequestBody Login Login) {
+    public String login(@RequestBody br.com.clsystem.CLSystem.model.entities.record.Login login) {
         try {
             // Cria o token de autenticação
             UsernamePasswordAuthenticationToken authenticationToken = 
-                new UsernamePasswordAuthenticationToken(Login.user(), Login.password());
+                new UsernamePasswordAuthenticationToken(login.user(), login.password());
 
             // Autentica o token
             Authentication authentication = authenticationManager.authenticate(authenticationToken);
